app/pyproject.toml
[project]
name = "collabai-app"
version = "0.1.0"
description = "collaboration-ai backend — hof-engine application implementing event-sourced team chat with first-class agents."
requires-python = ">=3.12"
dependencies = [
    "hof-engine>=0.5.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "asyncpg>=0.30",
    "psycopg[binary]>=3.2",
    "redis>=5.2",
    "celery[redis]>=5.4",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "argon2-cffi>=23.1",
    "python-jose[cryptography]>=3.3",
    "boto3>=1.35",
    "pillow>=11.0",
    "bleach>=6.2",
    "markdown-it-py>=3.0",
    "pywebpush>=2.0",
    "mcp>=1.0",
    "websockets>=14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.7",
    "mypy>=1.13",
    "freezegun>=1.5",
]

[project.scripts]
collabai-mcp = "domain.agent_api.bin:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["domain", "tables", "functions", "flows", "cron"]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = [".venv", "__pycache__", "node_modules"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]
